requirements.txt
pycryptodome == 3.18.0
PyQt5 == 5.15.9
Pyqt5-sip == 12.12.1
Pyqt5-Qt5 == 5.15.2
cryptography == 41.0.1
